backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
bcrypt==3.2.2
python-jose[cryptography]
python-multipart
pydantic